+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 
+- refactor: use `hyper` & `tower` instead of `reqwest` for feeder client
 - fix: contract 0 state diff fixed
 - refactor(rpc): re-worked rpc tower server and added proper websocket support
 - fix(network): added the FGW and gateway url to the chain config